requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.0
django-modeltranslation==0.18.9
Pillow==9.5.0
python-decouple==3.8
sqlparse==0.4.4
typing-extensions==4.5.0
whitenoise==6.4.0
